Add a FAQ for existing images still being served on local path
I saw that some ppl asked the same question than me, adding this in the FAQ will save you some times.

Also, it'd be nice if the plugins offers a solution to process existing remote images. It will allow ppl to use your plugin not only in new installations, but also in previous and old ones.